unsigned

+++ Basics/binary_number/binary1.cpp
@@ -14,6 +14,6 @@ using namespace std;
 
 int main() {
     cout << sizeof(int) << endl;
-    cout << sizeof(short int) << endl;
+    cout << sizeof(unsigned int) << endl;
     return 0;
 }
